requirements.txt
dbt-bigquery>=1.3.0
google-auth
google-cloud-bigquery
google-cloud-storage
google-cloud-aiplatform==1.18.3
google-cloud-pipeline-components==1.0.27
kfp==1.8.14
tdqm